awesome/rewrote tasklist widget from scratch
